refactor: 쿠키 접근 후 토큰 분기처리 추가

+++ src/app/_api/bookmark.ts
@@ -1,3 +1,4 @@
+import { authCheck } from "@/utils/function/authCheck";
 import { BookMarkedAddResponse } from "@/utils/types/bookmark/add";
 import { BookMarkedAllCheckResponse } from "@/utils/types/bookmark/allCheck";
 import { BookMarkedDeleteResponse } from "@/utils/types/bookmark/delete";
@@ -9,12 +10,16 @@ export type getCheckBookmarkResponse = { isBookmarked: boolean };
 //TODO: 토큰 없애기
 export const getCheckBookmark = async ({
   auctionId
-}: getCheckBookmarkParams): Promise<getCheckBookmarkResponse> => {
+}: getCheckBookmarkParams): Promise<getCheckBookmarkResponse | null> => {
+  const isTokenValid = authCheck();
+
+  if (!isTokenValid) return null;
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/bookmarks/${auctionId}`,
     {
       headers: {
-        Authorization: "Bearer "
+        Authorization: `Bearer ${isTokenValid}`
       }
     }
   );
@@ -27,13 +32,17 @@ export const getCheckBookmark = async ({
 
 export const addBookmark = async (
   auctionId: number
-): Promise<BookMarkedAddResponse> => {
+): Promise<BookMarkedAddResponse | null> => {
+  const isTokenValid = authCheck();
+
+  if (!isTokenValid) throw new Error("401");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/bookmarks/${auctionId}`,
     {
       method: "POST",
       headers: {
-        Authorization: "Bearer "
+        Authorization: `Bearer ${isTokenValid}`
       }
     }
   );
@@ -47,13 +56,17 @@ export const addBookmark = async (
 
 export const deleteBookmark = async (
   auctionId: number
-): Promise<BookMarkedDeleteResponse> => {
+): Promise<BookMarkedDeleteResponse | null> => {
+  const isTokenValid = authCheck();
+
+  if (!isTokenValid) throw new Error("401");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/bookmarks/${auctionId}`,
     {
       method: "DELETE",
       headers: {
-        Authorization: "Bearer "
+        Authorization: `Bearer ${isTokenValid}`
       }
     }
   );
@@ -66,12 +79,16 @@ export const deleteBookmark = async (
 };
 
 export const getCheckBookmarkList =
-  async (): Promise<BookMarkedAllCheckResponse> => {
+  async (): Promise<BookMarkedAllCheckResponse | null> => {
+    const isTokenValid = authCheck();
+
+    if (!isTokenValid) return null;
+
     const res = await fetch(
       `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/bookmarks`,
       {
         headers: {
-          Authorization: "Bearer "
+          Authorization: `Bearer ${isTokenValid}`
         }
       }
     );
